+++ src/app/app_settings.cc
@@ -4,6 +4,7 @@
 
 #include <nlohmann/json.hpp>
 
+
 namespace petri::app {
 
 using json = nlohmann::json;