+++ src/main.cc
@@ -3,6 +3,7 @@
 #include <iostream>
 #include <string>
 
+
 int main(int argc, char* argv[]) {
     if (argc < 2) {
         std::cerr << "Use config path as first cl arg\n";